Move tomography fitter data fn to fitters
